06-lab-complete/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.2
redis
python-multipart==0.0.12
psutil==6.0.0

# Agent Dependencies (Pip will solve compatibility)
langchain-google-genai
langgraph
langchain-core
langgraph-checkpoint-redis
tavily-python
python-dotenv
requests
